requirements.txt
PyYAML==5.2
urllib3==1.24.1
elasticsearch==7.1.0